backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis & Celery
redis==5.0.1
celery==5.3.6

# AI & ML
numpy<2.0
google-generativeai==0.3.2
groq==0.4.2
langdetect==1.0.9
sentence-transformers==2.3.1
chromadb==0.4.22

# Web Scraping
feedparser==6.0.10
playwright==1.41.0
playwright-stealth==1.0.6
beautifulsoup4==4.12.3
lxml==5.1.0

# Image Processing
Pillow==10.2.0
httpx==0.26.0
aiohttp==3.9.3

# Security
cryptography==42.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
tenacity==8.2.3
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
